Summary: minor, get rid of the libdatrie dependency Keywords: trie
This changeset eradicates the need for a specially patched libdatrie.  In the
long run, the datrie implementation will be replaced by a more suitable
and optimised trie implementation.  Certainly so before the 0.2 release.

The datrie library used here does not support alpha maps, and comes with a
special in-memory reader/writer which is used to (de)serialise the trie from the
back of an ute file.

Signed-off-by: Sebastian Freundt <sebastian.freundt@ga-group.nl>